Moved `MovieRatingController` to API routes and added OpenAPI specs and HTTP tests
