+++ tests/rest/api/movie-rating.http
@@ -1,0 +1,18 @@
+@tmdbId = 329
+GET http://127.0.0.1/api/fetchMovieRatingByTmdbdId?tmdbId={{tmdbId}}
+Accept: */*
+Cache-Control: no-cache
+Content-Type: application/json
+X-Auth-Token: {{xAuthToken}}
+
+###
+
+POST http://127.0.0.1/api/users/{{username}}/movies/1/rating
+Accept: */*
+Cache-Control: no-cache
+Content-Type: application/json
+X-Auth-Token: {{xAuthToken}}
+
+{
+  "rating": 10
+}
